+++ src/app/service/poap/DistributePOAP.ts
@@ -1,12 +1,14 @@
 import { AwaitMessagesOptions, DMChannel, GuildMember, Message, MessageAttachment } from 'discord.js';
 import axios from 'axios';
-import POAPUtils, { POAPFileParticipant } from '../../utils/POAPUtils';
+import POAPUtils, { FailedPOAPAttendee, POAPFileParticipant } from '../../utils/POAPUtils';
 import ValidationError from '../../errors/ValidationError';
 import { Db } from 'mongodb';
 import dbInstance from '../../utils/dbUtils';
 import constants from '../constants/constants';
 import { CommandContext } from 'slash-create';
 import { LogUtils } from '../../utils/Log';
+import { Buffer } from 'buffer';
+import { getBufferForFailedParticipants } from './EndPOAP';
 
 export default async (ctx: CommandContext, guildMember: GuildMember, event?: string): Promise<any> => {
 	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_DEGEN);
@@ -16,7 +18,21 @@ export default async (ctx: CommandContext, guildMember: GuildMember, event?: str
 	const participantsList: POAPFileParticipant[] = await askForParticipantsList(guildMember);
 	await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
 	const linksMessageAttachment: MessageAttachment = await askForLinksMessageAttachment(guildMember);
-	await POAPUtils.sendOutPOAPLinks(guildMember, participantsList, linksMessageAttachment, event);
+	const failedPOAPsList: FailedPOAPAttendee[] = await POAPUtils.sendOutPOAPLinks(guildMember, participantsList, linksMessageAttachment, event);
+	const failedPOAPsBuffer: Buffer = getBufferForFailedParticipants(failedPOAPsList);
+	await guildMember.send({
+		embeds: [
+			{
+				title: 'POAPs Distribution Results',
+				fields: [
+					{ name: 'Attempted to Send', value: `${participantsList.length}`, inline: true },
+					{ name: 'Successfully Sent', value: `${participantsList.length - failedPOAPsList.length}`, inline: true },
+					{ name: 'Failed to Send', value: `${failedPOAPsList.length}`, inline: true },
+				],
+			},
+		],
+		files: [{ name: 'failed_to_send_poaps.csv', attachment: failedPOAPsBuffer }],
+	});
 };
 
 export const askForParticipantsList = async (guildMember: GuildMember): Promise<POAPFileParticipant[]> => {