+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 3. Add option to automatically end poap events
    - format poap start and end messages
    - fix guest pass auto removal
+4. Return csv of failed participants at end of event
 
 ## 1.7.2-RELEASE (2021-10-12)